Add utils file; split main class in different controllers
