+++ backend/src/config/utils.js
@@ -1,0 +1,19 @@
+var API_KEY = '1f54bd990f1cdfb230adb312546d765d';
+var URL_BASE = 'https://api.themoviedb.org/3/movie/';
+var URL_SEARCH = 'https://api.themoviedb.org/3/search/movie';
+var URL_BASE_IMAGE_POSTER = 'https://image.tmdb.org/t/p/w342/';
+var URL_BASE_IMAGE_BACKDROP = 'https://image.tmdb.org/t/p/w300/';
+var URL_BASE_GENRE = 'https://api.themoviedb.org/3/genre/movie/list';
+var LANGUAGE = 'es-US';
+var API_VERSION = 'v1';
+
+module.exports = {
+    API_KEY,
+    URL_BASE,
+    URL_SEARCH,
+    URL_BASE_IMAGE_POSTER,
+    URL_BASE_IMAGE_BACKDROP,
+    URL_BASE_GENRE,
+    LANGUAGE,
+    API_VERSION
+}
